Fix migration: cast to numeric for ROUND with precision arg
PostgreSQL's ROUND(double precision, integer) doesn't exist — the
two-argument form only accepts numeric. Add ::numeric cast.

+++ drizzle/0039_backfill_sleep_efficiency.sql
@@ -15,7 +15,7 @@ WHERE efficiency_pct IS NULL
 -- 2. Fix fractional efficiency values from Apple Health XML import (stored as 0-1 instead of 0-100)
 -- Any efficiency_pct <= 1.0 is almost certainly a fraction that needs rescaling
 UPDATE fitness.sleep_session
-SET efficiency_pct = ROUND(efficiency_pct * 100, 1)
+SET efficiency_pct = ROUND((efficiency_pct * 100)::numeric, 1)
 WHERE efficiency_pct > 0
   AND efficiency_pct <= 1.0;
